apicommerce/requirements.txt
Django
pymysql
python-dotenv
Pillow
djangorestframework
djangorestframework-simplejwt
django-cors-headers